requirements.txt
requests
pandas
python-dotenv
openai
tqdm